Detecting update conflict for new file
If there is a new version in the remote repo and the local repo has a
new file with the same name, the local file is protected by renaming and
the remote file is downloaded.

Also switched to returning NULL for missing files in local sync table
and local file sys when getting their timestamp.